Return false from Resource#destroy if the resource is new.

+++ lib/dm-core/resource.rb
@@ -265,6 +265,7 @@ module DataMapper
     # --
     # @api public
     def destroy
+      return false if new_record?
       return false unless repository.delete(to_query)
 
       @new_record = true
